fix: Preserve fractional UV index
The /device/values endpoint returns UV index in standard units and can be
fractional (e.g. 0.9). Parsing it with _to_int truncated values below 1 to 0.

- CurrentConditions.uv_index is now float
- get_current_conditions parses uvi with _to_float
- Document uvi as number in the OpenAPI spec

+++ src/weathercloud/models.py
@@ -46,7 +46,7 @@ class CurrentConditions:
     rain_rate: float | None        # mm/h
     rain: float | None             # mm total
     solar_radiation: float | None  # W/m²
-    uv_index: int | None
+    uv_index: float | None         # standard units; can be fractional
     inside_temperature: float | None  # °C
     inside_humidity: int | None       # %
     inside_heat_index: float | None   # °C
